- Fixed license headers - Separated out LocalNeuralNet, TensorTwoD, TensorOneD classes - Fixed array indexing error in LocalNeuralNet
