+++ tutorial/ffnet_agent/Makefile
@@ -1,33 +1,6 @@
 #
-#  Copyright (c) 2015 - 2021, Intel Corporation
-#
-#  Redistribution and use in source and binary forms, with or without
-#  modification, are permitted provided that the following conditions
-#  are met:
-#
-#      * Redistributions of source code must retain the above copyright
-#        notice, this list of conditions and the following disclaimer.
-#
-#      * Redistributions in binary form must reproduce the above copyright
-#        notice, this list of conditions and the following disclaimer in
-#        the documentation and/or other materials provided with the
-#        distribution.
-#
-#      * Neither the name of Intel Corporation nor the names of its
-#        contributors may be used to endorse or promote products derived
-#        from this software without specific prior written permission.
-#
-#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-#  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-#  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-#  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-#  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-#  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-#  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-#  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-#  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-#  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY LOG OF THE USE
-#  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#  Copyright (c) 2015 - 2022, Intel Corporation
+#  SPDX-License-Identifier: BSD-3-Clause
 #
 
 TUTORIAL_LIB = libgeopmagent_cpu_ffnet_agent.so.0.0.0 \
@@ -53,14 +26,20 @@ EXTRA_CPPFLAGS:=$(EXTRA_CPPFLAGS)
 CXXFLAGS=$(GEOPM_CFLAGS) $(OMP_FLAGS)
 LDFLAGS=$(GEOPM_LDFLAGS) $(OMP_FLAGS)
 
-libgeopmagent_cpu_ffnet_agent.so.0.0.0: CPUFFNetAgent.cpp CPUFFNetAgent.hpp localnet.o
-	$(CXX) $(CXXFLAGS) $(LDFLAGS) $(GEOPM_PLUGIN_FLAGS) $(EXTRA_CPPFLAGS) CPUFFNetAgent.cpp localnet.o -o $@
+libgeopmagent_cpu_ffnet_agent.so.0.0.0: CPUFFNetAgent.cpp CPUFFNetAgent.hpp LocalNeuralNet.o TensorOneD.o TensorTwoD.o
+	$(CXX) $(CXXFLAGS) $(LDFLAGS) $(GEOPM_PLUGIN_FLAGS) $(EXTRA_CPPFLAGS) CPUFFNetAgent.cpp LocalNeuralNet.o TensorOneD.o TensorTwoD.o -o $@
 
-localnet.o: localnet.cpp localnet.hpp
-	$(CXX) $(CXXFLAGS) $(GEOPM_PLUGIN_FLAGS) -c localnet.cpp -o $@
+TensorOneD.o: TensorOneD.cpp TensorOneD.hpp
+	$(CXX) $(CXXFLAGS) $(GEOPM_PLUGIN_FLAGS) -c TensorOneD.cpp -o $@
 
-test: test.cpp localnet.o
-	$(CXX) $(CXXFLAGS) localnet.cpp test.cpp -o $@
+TensorTwoD.o: TensorTwoD.cpp TensorTwoD.hpp TensorOneD.hpp
+	$(CXX) $(CXXFLAGS) $(GEOPM_PLUGIN_FLAGS) -c TensorTwoD.cpp -o $@
+
+LocalNeuralNet.o: LocalNeuralNet.cpp LocalNeuralNet.hpp TensorOneD.hpp TensorTwoD.hpp
+	$(CXX) $(CXXFLAGS) $(GEOPM_PLUGIN_FLAGS) -c LocalNeuralNet.cpp -o $@
+
+test: test.cpp LocalNeuralNet.o TensorOneD.o TensorTwoD.o
+	$(CXX) $(CXXFLAGS) LocalNeuralNet.o test.cpp TensorOneD.o TensorTwoD.o -o $@
 
 clean:
 	rm -f $(TUTORIAL_LIB) *.o test
